backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy
pandas
numpy
FinanceDataReader
pykrx
python-dotenv
aiofiles
httpx